ブラウザにjsonデータ表示&ディレクトリ構成変更&HomePage追加
