+++ src/App.js
@@ -1,12 +1,15 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import UserMicropostsList from './components/users/microposts/GetUserMicroposts.jsx';
+
+import HomePage from './components/page/HomePage';
+import UserMicropostsList from './components/page/GetUserMicroposts.jsx';
 
 const App = () => {
   return (
     <BrowserRouter>
       <Routes>
-        <Route exact path='/:userId/microposts' element={<UserMicropostsList />}/>
+        <Route path='/' element={<HomePage />}/>
+        <Route path='/:userId/microposts' element={<UserMicropostsList />}/>
       </Routes>
     </BrowserRouter>
   );
