+++ src/components/page/HomePage.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+const HomePage = () => {
+    return (
+      <div>
+        <label>ユーザーIDを入力</label>
+        <input
+          type="number"
+        />
+      </div>
+    );
+  }
+
+export default HomePage;
